Added ideal position and velocity output class
Inside of the motion_forward() command, a new object will be created
that will output what the current speed and location should be.

The class assumes that the maximum speed is faster or very close to the
current speed.  It also assumes that the input distance was positive
because it starts with a positive acceleration and ends with a negative
acceleration value.

Class also assumes that the top speed is within the range of the
motors, and the robot can accelerate at the max acceleration up to the
max speed.  While operating with 2 motors these values will be limited