Added rotation function
Added the rotation function to motion.cpp.  sensors_range.cpp was
started but is not yet functional.  conf.h was adjusted to accommodate
both rotation and straight motions.

The encoders are now only reset at the end of motion functions instead
of the beginning of the functions.  This way overshoot after finishing
a motion but before starting the next motion is not lost, which would
increase error