Made motion parameters adjustable
Motion control parameters, such as max acceleration and velocity, are
now adjustable on the fly using added functions in motion.cpp.  The max
acceleration, deceleration, and velocity are all specific to one kind
of motion profile and can be adjusted anytime between motions but not
during.

The rangefinder equation was also fixed from last time