Removed unnecessary variables in table generation
Code is untested but compiles

Removed some of the unnecessary input parameters in the SweptTurnTable
class.

Added new constants to produce 45, 90, 135, and 180 degree turns.

Changed speed scaling to time scaling because it is more intuitive.

Still needs a switch case to determine the correct lookup table to use.

Memory use should be evaluated for our lookup table generation process.
 I think the current process could still be more efficient.  I don’t
think we need to be saving a table of velocities anymore.

The lookups could be smaller, and also could be too small depending on
the duration of the turn.  The table must be at least 1000*.5*turn
duration.  so if using a standard table size, the table length should
be based on a 180 degree turn.