Synchronize with mazesim-rewrite-20160329
Fixes the zero path size bug in the Driver. Removes an include from
the Navigator that breaks portability.

Copied data.h outright. A few style changes in there, plus one
additional bug fix for indexing an array out of bounds.

Did not synchronize the RobotDriver because continuous search is messy
and likely to change soon.