Refactor Menu
Previous implementation was a single overarching Menu class. That
didn't gain anything besides splitting up the code into more managable
chunks in functions.

A Menu object now handles a single string selection. A Menu is
constructed from an array of MenuItems, and then that Menu is called
like a function in order to "run" the Menu.

A MenuItem groups a display name with a function to be executed when
the MenuItem is selected in a Menu.

Deleted the code that _uses_ the Menu, to be added back in a later
commit.

Resolves #74.