Remove OPT -> SDIR back option
When the menu starts, the selected option does not default to the
active start direction. That's confusing to the user, since other
settings (such as OPT -> SPDS -> SVEL) default to the active value.

Scenario: The active value is East. The user goes to OPT -> SDIR and
sees NRTH. Believing the active value is NRTH, the user selects BACK.

As a quick solution, I removed the back option. This forces the user
to select one of the options so there is no question which value is
selected.

The proper fix is to add a way to specify a default value for a Menu.
Created issue #88 to address this.