Fix buffer overrun in Menu implementation
In the constructor, the length is calculated as
countLength(items) + 1, where the + 1 is for the appended BACK item.
When countLength(items) == kMaxLength, this results in an overrun.

Adjusted countLength() to return at most kMaxLength - 1. This keeps in
one place the complexity of deciding where to truncate an array that
is too long for a Menu's capacity.

Also, in the constructor, one item too many was being copied from the
items array. Practically, this had no ill effect on its own, because
that last "undefined" item was then overwritten by the BACK item.