Set up micromouse_main()
Though all sources are now in src/, Arduino still requires the main
"sketch" to be in the root of the project and to have the same name as
the project directory.

Created a micromouse.ino file whose job is to call micromouse_main(),
which we will treat as our "main" function. What used to be
micromouse.ino is now main.cpp.

This allows us to have a "main" file, with a "main" function, without
paying mind to Arduino's cute setup() and loop() construct. Also
relieves us of cute restrictions on where our main function can be, as
long as the root-level micromouse.ino is in tact.